add include/picrin/private dir
