+++ contrib/10.callcc/callcc.c
@@ -1,6 +1,6 @@
 #include "picrin.h"
-#include "picrin/object.h"
-#include "picrin/state.h"
+#include "picrin/private/object.h"
+#include "picrin/private/state.h"
 
 struct pic_fullcont {
   jmp_buf jmp;
