+++ extlib/benz/dict.c
@@ -4,7 +4,7 @@
 
 #include "picrin.h"
 #include "picrin/extra.h"
-#include "picrin/object.h"
+#include "picrin/private/object.h"
 
 KHASH_DEFINE(dict, pic_sym *, pic_value, kh_ptr_hash_func, kh_ptr_hash_equal)
