+++ extlib/benz/error.c
@@ -4,8 +4,8 @@
 
 #include "picrin.h"
 #include "picrin/extra.h"
-#include "picrin/object.h"
-#include "picrin/state.h"
+#include "picrin/private/object.h"
+#include "picrin/private/state.h"
 
 void
 pic_panic(pic_state PIC_UNUSED(*pic), const char *msg)
