+++ extlib/benz/file.c
@@ -3,7 +3,7 @@
  */
 
 #include "picrin.h"
-#include "picrin/state.h"
+#include "picrin/private/state.h"
 
 #ifndef EOF
 # define EOF (-1)
