+++ extlib/benz/string.c
@@ -4,7 +4,7 @@
 
 #include "picrin.h"
 #include "picrin/extra.h"
-#include "picrin/object.h"
+#include "picrin/private/object.h"
 
 struct pic_chunk {
   char *str;
