+++ extlib/benz/vector.c
@@ -4,7 +4,7 @@
 
 #include "picrin.h"
 #include "picrin/extra.h"
-#include "picrin/object.h"
+#include "picrin/private/object.h"
 
 pic_value
 pic_make_vec(pic_state *pic, int len, pic_value *argv)
